feat: move augment files for better organization
